値クラスとページクラスをそれぞれファイルに切り出す
